Instead of Input, I created InputRunner, which is forked off the main process and keeps fgets(STDIN) from blocking the parent.
This was my original use case which prompted me to look into PCNTL. I thought
that SIGALRM could break fgets(STDIN) as well, but I suppose that this doesn't
work as the process cannot send SIGSTOP and SIGCONT to the user's brain. When
dealing with other streams, it is possibly able to halt the processing for a
short time.