Changed the concept of how a process gets ended
The first approach was to call terminate and then have the Process return
false on loop as soon as it is done. This is error prone in a way that some
kind of "is active or not" logic has to be added to every implementation of
Timeshared, which also means that some additional code needs to be run on
every call of loop, which is useless most of the time.
So let terminate() decide instead if the program is ready to terminate or
not, or a combination of either loop or terminate, depending on the use
case.

The Counter class with a $modulo value is a good example: it either stops
when the max value or when the next multiple of $modulo is reached.
Depending on which one is reached first, a false value returned by loop or
a true value returned by terminate will end the process.

By now unresponsive processes can block indefinitely, but this will be
addressed later.