Added timeout for Timeshare::terminate.
As soon as Timeshare::terminate() is called, loop checks for the time passed
since terminate was called for the first time. If the timeout is surpassed,
Timeshare::kill() is called, which in term calls kill on any remaining
processes.

It has to be noted that this cannot work wonders as PHP with Timeshare is still
single threaded, so if a process actually blocks, it can't be killed.

A possible solution would be to have a lightweight main process from which a
worker process that runs Timeshare is forked off.