TaskEnvelope is a small wrapper around every instance of Timeshared.
This is supposed to encapsulate the state of tasks, so it is not necessary
to keep it in the scheduler. A first example is that the decision whether
it is necessary to run __tsStart is offloaded to TaskEnvelope::__tsLoop so
that the Scheduler does not have to care about Timeshared->__tsStart
altohether.