Timeshare uses TaskEnvelope to offload some of its functions.
Currently TaskEnvelope->__tsLoop() will call Timeshared->__tsStart() on first
call and Timeshared->tsLoop() on every subsequent call.

I made remove public because it might be a necessity at some point to move a
task from one scheduler to another. But while I am considering this, it is
possibly not the best solution to move the task, but the TaskEnvelope (so the
internal state of TaskEnvelope is kept).