Tucked away terminate & timeout in TaskEnvelope
Terminating processes is now "neatly" tucked away in TaskEnvelope. This is neat
on the one hand, as the terminated state does not need to be tracked in
Timeshare itself - something that will be necessary if I add functionality to
terminate single processes to Timeshare.
Less neat because I don't like bidirectional dependencies and need to have a
public Timeshare::remove for now, which should never be called from the outside.

There are solutions to this, however.

TimeshareObserverTest had to be changed: the prior implementation was
'synchronized' in that Timeshare::__tsTerminate() terminated all processes at
once, as long as they were not deferred.